repair double passing

+++ tvlm/il/il_builder.h
@@ -139,7 +139,7 @@ namespace tvlm {
         Instruction * getStringLiteral(std::string const & lit, const tiny::ASTBase * ast){
             auto i = stringLiterals_.find(lit);
             if (i == stringLiterals_.end()) {
-                auto arrSize = add(new LoadImm(lit.size() + 1, ast));
+                auto arrSize = add(new LoadImm((int64_t)lit.size() + 1, ast));
                 auto strType = registerType(new Type::Array(registerType(new Type::Char()), arrSize));
                 Instruction * addr = globalAdd(new tvlm::AllocG{strType, arrSize , ast});
                 stringLiterals_.insert(std::make_pair(lit, addr));
